Add to results to merging list

add.res.to.merging.list Adds EWCE results to a list for merging analysis

add.res.to.merging.list(full_res, existing_results = NULL)

Arguments

full_res
results list generated using bootstrap.enrichment.test or ewce_expression_data functions. Multiple results tables can be merged into one results table, as long as the 'list' column is set to distinguish them.

existing_results
Output of previous rounds from adding results to list. Leave empty if this is the first item in the list.

Value

merged results list

Examples

# Load the single cell data
data("ctd")

# Load the data
data(tt_alzh)
data(tt_alzh_BA36)
data(tt_alzh_BA44)

# Run EWCE analysis
tt_results = ewce_expression_data(sct_data=ctd,tt=tt_alzh,annotLevel=1,
     ttSpecies="human",sctSpecies="mouse")

#>   [1] "Myom1"         "Rbm17"         "Max"           "Mbd3"         
#>   [5] "Wdr33"         "Acss1"         "Pcyt1b"        "Bcam"         
#>   [9] "Prpf38b"       "Ppp2r5c"       "Ica1"          "Cobl"         
#>  [13] "Marveld1"      "Gemin4"        "Zfp668"        "Casc3"        
#>  [17] "Aff3"          "Tcf3"          "Taf1c"         "Gna12"        
#>  [21] "Lamp2"         "Inpp5d"        "Phf14"         "Stk11"        
#>  [25] "Plxnd1"        "Aebp1"         "Pgf"           "5430427O19Rik"
#>  [29] "Gad1"          "App"           "Casp9"         "Mid1ip1"      
#>  [33] "Cep70"         "Dnajc1"        "Cbx5"          "Llph"         
#>  [37] "Rbm6"          "Tbc1d2b"       "Csk"           "Col6a1"       
#>  [41] "1110034G24Rik" "Vars2"         "Lama2"         "Aup1"         
#>  [45] "Ctbp2"         "Atg3"          "Aifm3"         "Gga1"         
#>  [49] "Zcchc24"       "Prtg"          "Ralgps1"       "Ncan"         
#>  [53] "Otud4"         "Phf21a"        "Adra2c"        "Iqcc"         
#>  [57] "Ptbp1"         "Col27a1"       "Zfp526"        "Nrp2"         
#>  [61] "Flnb"          "Dcn"           "Zfp36l1"       "Antxr1"       
#>  [65] "Plod3"         "Ddit4l"        "Hnrnph3"       "Apba3"        
#>  [69] "Taf5"          "Mmd2"          "Kif1c"         "Zswim1"       
#>  [73] "Myh14"         "Wwox"          "Tmem47"        "Zfp36l2"      
#>  [77] "Palld"         "Mfge8"         "Nfic"          "Ilf3"         
#>  [81] "Armc3"         "Myh11"         "Ppp4r1"        "Plekha8"      
#>  [85] "Gas2l1"        "Tal1"          "Atp8b2"        "Slc40a1"      
#>  [89] "Enah"          "1110059E24Rik" "Nsd1"          "Rprm"         
#>  [93] "Golim4"        "Paox"          "Numa1"         "Rabep1"       
#>  [97] "Tcf12"         "Fastk"         "Cnot4"         "Cxcr4"        
#> [101] "Rrad"          "Dmc1"          "Rassf9"        "Abca9"        
#> [105] "Hbp1"          "Klk7"          "Nedd4"         "Adck1"        
#> [109] "Smarcc1"       "Ezr"           "Gpr161"        "Rapgef3"      
#> [113] "Snrnp48"       "Abi3bp"        "Nav1"          "Ssr1"         
#> [117] "Pknox1"        "Qk"            "Osbp2"         "Col5a3"       
#> [121] "Fgfr1"         "Repin1"        "Dlgap4"        "Atp5sl"       
#> [125] "Rimkla"        "Pdcd2"         "Alcam"         "P2rx7"        
#> [129] "Smyd4"         "Mtss1l"        "Sfxn5"         "Unc5b"        
#> [133] "Ddc"           "Pcdh1"         "Igf2"          "Sec11a"       
#> [137] "Osbpl2"        "3110079O15Rik" "Sall3"         "Git1"         
#> [141] "Llgl2"         "Zfp385a"       "Kank2"         "Rps21"        
#> [145] "Smg5"          "Gnao1"         "Csnk1d"        "Ubxn1"        
#> [149] "Sos1"          "Ascl1"         "Map2k3"        "Nf2"          
#> [153] "Fbxw4"         "Ocln"          "Zc3h14"        "S100a4"       
#> [157] "Rftn2"         "Zfp444"        "Rab28"         "Notch2"       
#> [161] "Efcab2"        "Maf"           "Hes5"          "Chdh"         
#> [165] "Mdm2"          "Arhgap22"      "Rnf11"         "Crtap"        
#> [169] "Frzb"          "Gja1"          "Lmna"          "Rheb"         
#> [173] "Tmem184b"      "Shroom1"       "Ahnak"         "Itsn1"        
#> [177] "Jmjd6"         "Kcnj10"        "Bgn"           "Atg12"        
#> [181] "Ngfr"          "Slc25a29"      "Pcmtd1"        "Adam33"       
#> [185] "Rnd2"          "5031439G07Rik" "Mif4gd"        "Atp6v0e"      
#> [189] "Hist3h2a"      "Tmcc2"         "Mrgprf"        "Nrip2"        
#> [193] "Lef1"          "Sox12"         "Med25"         "Jund"         
#> [197] "Acp6"          "Ak1"           "Cd79b"         "Wwc1"         
#> [201] "Itga7"         "Map3k11"       "Cspg4"         "Vat1"         
#> [205] "Zfp692"        "Selplg"        "Dbi"           "Pou3f2"       
#> [209] "Doc2b"         "Itgb5"         "Polr3g"        "Pacs1"        
#> [1] "astrocytes_ependymal"
#> [1] 0
#> [1] "Fold enrichment: 1.23668798716229"
#> [1] "Standard deviations from mean: 3.94674630618526"
#> [1] ""
#> [1] "endothelial-mural"
#> [1] 0
#> [1] "Fold enrichment: 1.25609384486044"
#> [1] "Standard deviations from mean: 3.65838310917809"
#> [1] ""
#> [1] "interneurons"
#> [1] 1
#> [1] ""
#> [1] "microglia"
#> [1] 0
#> [1] "Fold enrichment: 1.20388311697137"
#> [1] "Standard deviations from mean: 2.76646602722591"
#> [1] ""
#> [1] "oligodendrocytes"
#> [1] 0.81
#> [1] ""
#> [1] "pyramidal CA1"
#> [1] 1
#> [1] ""
#> [1] "pyramidal SS"
#> [1] 1
#> [1] ""
#>   [1] "Prmt5"         "Atg7"          "Ivd"           "Ica1"         
#>   [5] "Farsb"         "Aifm1"         "Pole3"         "Psmd11"       
#>   [9] "Dpm2"          "Vrk3"          "Dnajc5"        "Gphn"         
#>  [13] "Tubb5"         "Ets2"          "Pygb"          "Fgf12"        
#>  [17] "Me2"           "Tfrc"          "Adcy3"         "Ramp3"        
#>  [21] "Cops7a"        "Snrpb"         "Retsat"        "Reep1"        
#>  [25] "Nap1l5"        "Rufy3"         "Psmg1"         "Dclk1"        
#>  [29] "Carf"          "Vsnl1"         "Sec23a"        "Liph"         
#>  [33] "Pias2"         "Kif9"          "Hpcal1"        "Med14"        
#>  [37] "Lmo4"          "Ank"           "Col12a1"       "Lrrc7"        
#>  [41] "Adam23"        "Lzic"          "Fgf14"         "Tollip"       
#>  [45] "Gnal"          "Lpin2"         "Rsrc1"         "Necap1"       
#>  [49] "Atp6v1b2"      "Kdm5c"         "Coq6"          "Rhbdd2"       
#>  [53] "Mef2c"         "Gabrd"         "Ajap1"         "Pcbp2"        
#>  [57] "Brwd1"         "Atp6v1g1"      "Mysm1"         "H2afy"        
#>  [61] "Parp2"         "Fam175b"       "Mad2l1bp"      "Aak1"         
#>  [65] "Uqcrc2"        "Rragb"         "Cd200"         "Phtf1"        
#>  [69] "Psd3"          "Rer1"          "Grin2a"        "Atp6ap2"      
#>  [73] "Slc27a2"       "Ddx19a"        "Cirbp"         "Camk1g"       
#>  [77] "Srp72"         "Zfp697"        "Cdc5l"         "Prkag2"       
#>  [81] "Pdk2"          "Gng3"          "Ddx10"         "Farsa"        
#>  [85] "Nrxn3"         "Hnrnpd"        "Pgs1"          "Fgfr1op"      
#>  [89] "Nrn1"          "Commd4"        "Msi2"          "Mapk10"       
#>  [93] "Npc1"          "Cap1"          "Pdss1"         "Eif4b"        
#>  [97] "Pgk1"          "Stx18"         "Slc20a1"       "Gprasp1"      
#> [101] "Atp5b"         "Ppp2ca"        "Zcchc17"       "Tubgcp4"      
#> [105] "Dlat"          "Ypel1"         "Rab2a"         "Gns"          
#> [109] "Snip1"         "Thumpd3"       "Pcyox1l"       "Pgm2l1"       
#> [113] "Gdap1"         "Nat10"         "Arih2"         "Prkab1"       
#> [117] "Rgs4"          "Asap1"         "Amph"          "Prep"         
#> [121] "Kcnq3"         "Sema3a"        "G6pc3"         "Wdr1"         
#> [125] "Nek3"          "Ndufa10"       "Vamp1"         "Prkcd"        
#> [129] "Fosb"          "Mtpap"         "Lgals8"        "Rnf41"        
#> [133] "Igf1"          "Krba1"         "Pfkm"          "Slc35a3"      
#> [137] "Stmn2"         "Scg3"          "Psma3"         "Snx17"        
#> [141] "Ephx2"         "Atp8a2"        "Rfc2"          "Crls1"        
#> [145] "Daglb"         "Gabra4"        "Dennd3"        "Jak3"         
#> [149] "Ogfod1"        "Pex1"          "Tmem59"        "Rbm3"         
#> [153] "Sgsh"          "Bcl2l13"       "Sdc3"          "Stat4"        
#> [157] "Mlx"           "Kctd5"         "Fads3"         "Trappc2l"     
#> [161] "Tbc1d9"        "Ptk2b"         "Thnsl2"        "Txn2"         
#> [165] "Mcph1"         "Snx10"         "Hs3st2"        "Camk2g"       
#> [169] "Npat"          "Kdelr2"        "Xpnpep1"       "Axin2"        
#> [173] "Mak16"         "Cacnb2"        "Prpf4"         "Zfp398"       
#> [177] "Wars"          "Cep57"         "Eif5"          "Ndfip1"       
#> [181] "Ndrg4"         "Nol6"          "Letmd1"        "Prss23"       
#> [185] "Gaa"           "Bicd2"         "Dhrs7b"        "M6pr"         
#> [189] "Usp11"         "Cdc37l1"       "Map2k5"        "Ywhaz"        
#> [193] "Fen1"          "Atp6ap1"       "Nedd4l"        "Mcam"         
#> [197] "Vps53"         "Slc17a7"       "Cobll1"        "Actr1a"       
#> [201] "Tiam2"         "Mrps18b"       "Slc38a7"       "Gabrg2"       
#> [205] "Pex3"          "Pdcd4"         "4930430F08Rik" "Lrrtm2"       
#> [209] "Cadps2"        "Mybbp1a"       "Ttc19"         "Lrrfip1"      
#> [213] "Susd4"         "Mat2b"         "Cyb561"        "Snap47"       
#> [217] "Sssca1"        "Tomm20"       
#> [1] "astrocytes_ependymal"
#> [1] 1
#> [1] ""
#> [1] "endothelial-mural"
#> [1] 1
#> [1] ""
#> [1] "interneurons"
#> [1] 0
#> [1] "Fold enrichment: 1.23535922151649"
#> [1] "Standard deviations from mean: 4.05832350691913"
#> [1] ""
#> [1] "microglia"
#> [1] 1
#> [1] ""
#> [1] "oligodendrocytes"
#> [1] 1
#> [1] ""
#> [1] "pyramidal CA1"
#> [1] 0
#> [1] "Fold enrichment: 1.19856421316632"
#> [1] "Standard deviations from mean: 4.20572948282647"
#> [1] ""
#> [1] "pyramidal SS"
#> [1] 0
#> [1] "Fold enrichment: 1.29635745327498"
#> [1] "Standard deviations from mean: 5.82227295327463"
#> [1] ""

tt_results_36 = ewce_expression_data(sct_data=ctd,tt=tt_alzh_BA36,annotLevel=1,
     ttSpecies="human",sctSpecies="mouse")

#>   [1] "Akap14"        "Arsb"          "Pigv"          "Zzz3"         
#>   [5] "Jtb"           "Slc39a1"       "Gpm6b"         "Heph"         
#>   [9] "Rps11"         "Cobl"          "Dlg5"          "Mmaa"         
#>  [13] "Vgll4"         "Pno1"          "Aga"           "Llgl1"        
#>  [17] "Sh3glb1"       "Arnt"          "Tfrc"          "Gna11"        
#>  [21] "Dpf3"          "Thbd"          "Rbm45"         "Rbm15b"       
#>  [25] "Acadvl"        "Fam107a"       "Ablim2"        "Brd7"         
#>  [29] "Ccdc77"        "Bard1"         "Smc6"          "Ripk1"        
#>  [33] "Gzf1"          "Slc35d1"       "Gprc5b"        "Dsg2"         
#>  [37] "Srprb"         "Prr14"         "Pibf1"         "Znrf2"        
#>  [41] "Hps3"          "Gabbr1"        "Slc6a19"       "Hkdc1"        
#>  [45] "Itih5"         "Ikzf4"         "Nkd1"          "Kcnn2"        
#>  [49] "Cdca7"         "Gpc4"          "Homer3"        "Fam149b"      
#>  [53] "Zbtb43"        "Slc7a11"       "Smchd1"        "Col27a1"      
#>  [57] "Elf2"          "Kctd7"         "Taok2"         "Setd3"        
#>  [61] "Gne"           "Ercc5"         "Efs"           "Ddit4l"       
#>  [65] "Rap2c"         "Sorbs1"        "Ncl"           "Klhl14"       
#>  [69] "Snta1"         "Vps37a"        "Gatm"          "Zfp36l2"      
#>  [73] "Rer1"          "Ubiad1"        "Htra3"         "Htra2"        
#>  [77] "Prickle1"      "Antxr2"        "Fcrls"         "Lgals1"       
#>  [81] "Ddx17"         "Masp1"         "Slc38a1"       "Lrig1"        
#>  [85] "Ccdc97"        "Nhsl1"         "Zc3h10"        "Plcd3"        
#>  [89] "Rabep1"        "Bag1"          "Slc48a1"       "Rab26"        
#>  [93] "Derl2"         "Sema5a"        "Mbd6"          "Itch"         
#>  [97] "Rrad"          "5730480H06Rik" "Ino80d"        "Wdr91"        
#> [101] "C330018D20Rik" "Acsbg1"        "Add3"          "Myo5c"        
#> [105] "Rapgef3"       "Kbtbd8"        "Wdr6"          "Gfap"         
#> [109] "Clu"           "Slc9a3r1"      "Snx5"          "Pik3c3"       
#> [113] "Plod1"         "Ddx52"         "Clk4"          "Art3"         
#> [117] "Fstl1"         "Litaf"         "Hk2"           "Plxnb2"       
#> [121] "Srd5a3"        "Apaf1"         "Grik1"         "Pttg1ip"      
#> [125] "Depdc1b"       "Osbpl2"        "Xpa"           "Tmem62"       
#> [129] "Sfrp1"         "Cthrc1"        "Tor1aip2"      "Ctsh"         
#> [133] "Cnot2"         "Zhx3"          "Smg5"          "Tbc1d16"      
#> [137] "Atr"           "Aplnr"         "Ubxn1"         "Snx1"         
#> [141] "Tpcn2"         "Irak2"         "Stx6"          "Cyth1"        
#> [145] "Rad18"         "Psph"          "Ahsa2"         "Bcl2l13"      
#> [149] "Notch2"        "Rcsd1"         "Mthfsd"        "Arhgap22"     
#> [153] "Trim35"        "Maml2"         "Vcan"          "Strap"        
#> [157] "Ppm1f"         "Pus10"         "Man1a2"        "Kdelr2"       
#> [161] "Thrap3"        "Tbc1d15"       "Nacc2"         "Sigirr"       
#> [165] "Fbxo4"         "Arhgef7"       "Ahnak"         "Cpne6"        
#> [169] "Arap1"         "Bgn"           "Ikbkb"         "Atp6v0a2"     
#> [173] "Pbxip1"        "Ccdc138"       "Tmem5"         "Atp1a2"       
#> [177] "0610009O20Rik" "5031439G07Rik" "Cnot1"         "Trp53bp2"     
#> [181] "Nfatc2"        "Lrp1"          "Tmf1"          "Poldip3"      
#> [185] "Adat2"         "Sntb1"         "Ralgps2"       "Prlr"         
#> [189] "Ehd1"          "Ncor1"         "Itga7"         "Xylt2"        
#> [193] "Prcp"          "Tbc1d9b"       "Agt"           "Cog2"         
#> [197] "Slc38a11"      "Ncstn"         "Mtmr3"         "Stxbp4"       
#> [201] "Lims2"         "Ttc7"          "C330027C09Rik" "Mfsd8"        
#> [205] "Zfp148"        "Kdm6a"         "Irf5"          "Tmem121"      
#> [1] "astrocytes_ependymal"
#> [1] 0
#> [1] "Fold enrichment: 1.36936067008076"
#> [1] "Standard deviations from mean: 6.40446573399607"
#> [1] ""
#> [1] "endothelial-mural"
#> [1] 0.01
#> [1] "Fold enrichment: 1.10467435799122"
#> [1] "Standard deviations from mean: 1.88920992887237"
#> [1] ""
#> [1] "interneurons"
#> [1] 1
#> [1] ""
#> [1] "microglia"
#> [1] 0
#> [1] "Fold enrichment: 1.18658938241039"
#> [1] "Standard deviations from mean: 2.60584869081001"
#> [1] ""
#> [1] "oligodendrocytes"
#> [1] 0.21
#> [1] ""
#> [1] "pyramidal CA1"
#> [1] 1
#> [1] ""
#> [1] "pyramidal SS"
#> [1] 1
#> [1] ""
#>   [1] "Nfatc1"        "Tbc1d5"        "Zfp142"        "Crim1"        
#>   [5] "Nsun4"         "Atxn3"         "Ankrd17"       "Ntrk3"        
#>   [9] "Lmln"          "Usp33"         "Zfp641"        "Vgll4"        
#>  [13] "Dnajb12"       "Klf13"         "Ppm1k"         "Eml4"         
#>  [17] "Mknk1"         "Eps15l1"       "Secisbp2l"     "Trpv2"        
#>  [21] "Wbp4"          "Fam13a"        "Egfr"          "Gria2"        
#>  [25] "Rufy3"         "Klhl22"        "Tmem144"       "Rbbp6"        
#>  [29] "Frmd3"         "Sox4"          "Rbmx2"         "Col4a3bp"     
#>  [33] "Fam155a"       "Spsb1"         "Pdpk1"         "Pou2f1"       
#>  [37] "Rgs12"         "1700030K09Rik" "Dhx36"         "Fzd8"         
#>  [41] "Slco3a1"       "Rapgef2"       "Fbxl18"        "Coq9"         
#>  [45] "Ireb2"         "Ergic2"        "Brd2"          "Gtpbp8"       
#>  [49] "Rnf115"        "Rap2a"         "Zyg11b"        "Sesn2"        
#>  [53] "Insr"          "Grid2"         "Akap6"         "Edn1"         
#>  [57] "Zcrb1"         "Tcf25"         "Tmem53"        "Rspry1"       
#>  [61] "Neurog2"       "Stk4"          "Vcp"           "Erc1"         
#>  [65] "Trio"          "Smc5"          "Rc3h2"         "Thumpd1"      
#>  [69] "Rbm25"         "Klf9"          "Nadk"          "Slitrk2"      
#>  [73] "Znrd1"         "Emilin2"       "Sh3rf1"        "Tox4"         
#>  [77] "Shmt1"         "Psd3"          "Pold1"         "Armc9"        
#>  [81] "Snca"          "Klf7"          "Pcdh9"         "Mapt"         
#>  [85] "Plekha1"       "Ap3d1"         "Ccdc59"        "Cdk9"         
#>  [89] "Nrxn3"         "Trnau1ap"      "Zfp652"        "Bbx"          
#>  [93] "Cap1"          "Emx1"          "Anks1b"        "Kcnk3"        
#>  [97] "Slc7a1"        "Slc20a1"       "Zfp397"        "Adrb3"        
#> [101] "Larp4"         "Brd4"          "Tmem97"        "Ifnar2"       
#> [105] "Ttc3"          "Ldb3"          "Cenpl"         "Pik3r1"       
#> [109] "Smarca4"       "Oprk1"         "Sh2b2"         "Nav1"         
#> [113] "Xdh"           "Ak2"           "Zfp800"        "Arih2"        
#> [117] "Fnbp4"         "Fcrl1"         "Fam184b"       "Ppig"         
#> [121] "Foxk2"         "Ank3"          "Cog7"          "Kif5c"        
#> [125] "Sema6a"        "Rims2"         "Tcf4"          "Gpr155"       
#> [129] "Nktr"          "Wiz"           "Mbd4"          "Zfp90"        
#> [133] "Kcnb1"         "Mtrf1l"        "Mycbp2"        "Gnao1"        
#> [137] "Tpcn2"         "Frmpd4"        "Caml"          "Nrxn1"        
#> [141] "Sobp"          "Egr4"          "Srrm2"         "Irs2"         
#> [145] "Glg1"          "Maf"           "Ncam1"         "Sbf1"         
#> [149] "Med28"         "Anapc5"        "Ubn2"          "Hic2"         
#> [153] "Mcph1"         "Magi1"         "Abcc3"         "Srpk2"        
#> [157] "Epha4"         "Cda"           "Ash1l"         "Nacc2"        
#> [161] "Prpf40a"       "Ubn1"          "Chd7"          "Taf1d"        
#> [165] "Fbxw7"         "Eif5"          "Cep68"         "Lyar"         
#> [169] "Atf6"          "Mecp2"         "Sqle"          "Zfp207"       
#> [173] "Zfp609"        "Ecel1"         "Fbxo21"        "Kpnb1"        
#> [177] "Tmtc2"         "Ddx24"         "Hsf1"          "Ficd"         
#> [181] "Plxna2"        "Tulp4"         "Igf2r"         "Ankrd11"      
#> [185] "Nucks1"        "Wnt7b"         "Ttc17"         "Runx1t1"      
#> [189] "Slc25a37"      "Ccdc63"        "Kif1a"         "Slc1a2"       
#> [193] "Cd47"          "Arid1b"        "Pwp1"          "Pus3"         
#> [197] "Atp1b1"        "Lrrtm2"        "Rp9"           "Ddx6"         
#> [201] "Adam19"        "Mlec"          "Fnbp1"         "Mtmr3"        
#> [205] "Gas2l3"        "Rsf1"          "Kcnj11"        "Gnas"         
#> [209] "Tnpo3"         "Snx19"         "Nrp1"         
#> [1] "astrocytes_ependymal"
#> [1] 1
#> [1] ""
#> [1] "endothelial-mural"
#> [1] 0.92
#> [1] ""
#> [1] "interneurons"
#> [1] 0.37
#> [1] ""
#> [1] "microglia"
#> [1] 0.11
#> [1] ""
#> [1] "oligodendrocytes"
#> [1] 0.54
#> [1] ""
#> [1] "pyramidal CA1"
#> [1] 0.06
#> [1] ""
#> [1] "pyramidal SS"
#> [1] 0.11
#> [1] ""

tt_results_44 = ewce_expression_data(sct_data=ctd,tt=tt_alzh_BA44,annotLevel=1,
     ttSpecies="human",sctSpecies="mouse")

#>   [1] "Creb3"         "Tmem87a"       "Tyk2"          "Wdr31"        
#>   [5] "Pex5"          "Vwa1"          "Acss1"         "Prmt2"        
#>   [9] "Htatip2"       "Rgs10"         "Dock1"         "Dennd5a"      
#>  [13] "Dhx16"         "Mcf2l"         "Flii"          "Cd274"        
#>  [17] "Cobl"          "Casc3"         "Arpp19"        "Ppfibp2"      
#>  [21] "Mfsd10"        "Stk36"         "Gtpbp6"        "Pkdcc"        
#>  [25] "Gab1"          "Kcnip3"        "Rbbp9"         "Wac"          
#>  [29] "Ranbp3"        "Brd3"          "Klf16"         "Notch1"       
#>  [33] "Pld1"          "Zfp445"        "Plxnd1"        "Zfpm1"        
#>  [37] "Aspa"          "Mid1ip1"       "Rps24"         "Bcan"         
#>  [41] "Arhgef10l"     "Gzf1"          "Mov10"         "Ralbp1"       
#>  [45] "Fzd8"          "Zfp689"        "Vars2"         "Acin1"        
#>  [49] "Rnf220"        "Ptp4a2"        "Insr"          "Pard6a"       
#>  [53] "Homer3"        "Pdlim2"        "Pld4"          "Ddx49"        
#>  [57] "Ctcf"          "Wdr53"         "Taok2"         "Tekt1"        
#>  [61] "Sap18"         "Wwox"          "Patz1"         "Ddx23"        
#>  [65] "Stac3"         "Zfp36l2"       "Pvr"           "Dhx38"        
#>  [69] "Cstb"          "Tns3"          "Cbr4"          "Map4k4"       
#>  [73] "Ppp4r1"        "Limch1"        "Arhgef2"       "Stard7"       
#>  [77] "Wnk1"          "Fam171a1"      "Cd82"          "Mchr1"        
#>  [81] "Spock3"        "Fcrls"         "Daxx"          "Tcof1"        
#>  [85] "Hnrnpab"       "Apbb2"         "Lpar1"         "Arrdc2"       
#>  [89] "Npc1"          "Tspan17"       "Baz1b"         "Bag1"         
#>  [93] "Kat5"          "Heca"          "Stk40"         "Cxcr4"        
#>  [97] "Fbxo18"        "Aspdh"         "Zfp637"        "Larp6"        
#> [101] "Thap11"        "Mylip"         "Smarcc1"       "Dusp10"       
#> [105] "Dlg1"          "Ezr"           "Rapgef3"       "Hexim1"       
#> [109] "Fbl"           "Gigyf2"        "Kbtbd4"        "Pomt2"        
#> [113] "Pum2"          "Cited4"        "Rnf207"        "Fuk"          
#> [117] "Ube4b"         "Sap30bp"       "Ankzf1"        "Gjb1"         
#> [121] "3110079O15Rik" "Git1"          "Ggt7"          "Fbxw8"        
#> [125] "Tbc1d16"       "Ndufv3"        "Hes6"          "Gsg1l"        
#> [129] "Ascl1"         "Ric8b"         "Osbpl7"        "Lrrc8d"       
#> [133] "Nf2"           "Ubac1"         "Fbxw4"         "Myo10"        
#> [137] "Ahsa2"         "Nfasc"         "Pusl1"         "Maf"          
#> [141] "Trabd"         "Mthfsd"        "Smarcd3"       "Ttll9"        
#> [145] "Arhgap22"      "Vcan"          "Safb2"         "Gpr108"       
#> [149] "Crtap"         "Qdpr"          "Acy3"          "Zfp174"       
#> [153] "Oraov1"        "Plekhb1"       "Spop"          "Nlgn3"        
#> [157] "Fam57b"        "Pkp4"          "Tmem184b"      "Exoc3"        
#> [161] "Shroom1"       "Tmcc3"         "Trim41"        "Ankrd13b"     
#> [165] "Fam173b"       "Mex3a"         "Rasgrp3"       "Ptpn11"       
#> [169] "Exosc3"        "Sqle"          "Gtf3c5"        "Slc38a2"      
#> [173] "Hs1bp3"        "Atp1b3"        "Dusp7"         "Sh3gl3"       
#> [177] "Tmem101"       "Aig1"          "Tmcc2"         "Nrip2"        
#> [181] "Bad"           "Nfkbib"        "Tmem119"       "Adamts13"     
#> [185] "Rnf130"        "Ppm1d"         "Cdr2l"         "Ncor1"        
#> [189] "Srrt"          "Mcm7"          "Vsig2"         "Cd84"         
#> [193] "Sccpdh"        "Selplg"        "Capn5"         "Itpk1"        
#> [197] "Rcor2"         "Pgpep1"        "R3hcc1"        "Chka"         
#> [201] "Vezf1"         "Nap1l1"        "Fbxo32"       
#> [1] "astrocytes_ependymal"
#> [1] 0.78
#> [1] ""
#> [1] "endothelial-mural"
#> [1] 0.64
#> [1] ""
#> [1] "interneurons"
#> [1] 1
#> [1] ""
#> [1] "microglia"
#> [1] 0.02
#> [1] "Fold enrichment: 1.14594979994608"
#> [1] "Standard deviations from mean: 2.21000300227331"
#> [1] ""
#> [1] "oligodendrocytes"
#> [1] 0
#> [1] "Fold enrichment: 1.42845891462859"
#> [1] "Standard deviations from mean: 6.87026171504374"
#> [1] ""
#> [1] "pyramidal CA1"
#> [1] 1
#> [1] ""
#> [1] "pyramidal SS"
#> [1] 0.97
#> [1] ""
#>   [1] "Lgals3"        "Rab36"         "Uba2"          "Mab21l2"      
#>   [5] "Dock7"         "Vegfc"         "Fez2"          "Stk17b"       
#>   [9] "Ica1"          "Bckdhb"        "Ntrk3"         "Nqo1"         
#>  [13] "Aim2"          "Dusp11"        "Ttc21b"        "Ahcyl1"       
#>  [17] "Arrdc3"        "Pno1"          "Zscan12"       "Sardh"        
#>  [21] "Ppm1k"         "Rpl35a"        "Eps15l1"       "Cdh8"         
#>  [25] "Stk38l"        "Pla2r1"        "Rragc"         "Tph1"         
#>  [29] "Cmpk2"         "Wwp2"          "Cpne8"         "Cdca3"        
#>  [33] "Slamf8"        "Det1"          "Mki67"         "Med31"        
#>  [37] "Fgf7"          "Tubb2b"        "Nup214"        "Cd38"         
#>  [41] "Fosl2"         "Mlf1"          "Ola1"          "Pigo"         
#>  [45] "Rarres1"       "Ttr"           "Wrap53"        "Srd5a1"       
#>  [49] "Gbp5"          "Ccdc68"        "Sv2c"          "Pdk1"         
#>  [53] "Adcyap1"       "Chmp1b"        "Gadd45a"       "Bicc1"        
#>  [57] "Otud4"         "Luzp2"         "Mrps30"        "Gls"          
#>  [61] "Pou3f3"        "Itga8"         "Dync2li1"      "Rlim"         
#>  [65] "Mmp24"         "Cd40"          "Fkbp14"        "Ide"          
#>  [69] "Setdb2"        "Abhd2"         "Il6st"         "Dnase1l3"     
#>  [73] "Cds1"          "Ptprg"         "Cp"            "Tuft1"        
#>  [77] "Lmcd1"         "Gabrr1"        "Rfesd"         "Slc35e1"      
#>  [81] "Dsc3"          "Ddx58"         "Hectd2"        "Psd3"         
#>  [85] "Cstf3"         "Sec22a"        "Tmtc1"         "Pdia3"        
#>  [89] "Dph5"          "Car12"         "Atl2"          "Efemp1"       
#>  [93] "Mchr1"         "Atp5g2"        "Smad3"         "Hnrnpd"       
#>  [97] "Slc22a4"       "Aste1"         "Slc16a6"       "Cenpj"        
#> [101] "Tmem97"        "Phtf2"         "Fmnl3"         "Nat10"        
#> [105] "Hsf4"          "Agfg1"         "Rnh1"          "Mllt3"        
#> [109] "Bcap29"        "Dnajc18"       "Prkab1"        "Hmgcl"        
#> [113] "Nid1"          "Stx17"         "Fbxw2"         "Nudcd1"       
#> [117] "Hddc2"         "Kcnq3"         "Smyd4"         "Ppp1r3d"      
#> [121] "Spcs2"         "Lgals8"        "Ddx31"         "Ppp2r5e"      
#> [125] "Dtna"          "Sf1"           "Tnfaip2"       "Rfc5"         
#> [129] "8430408G22Rik" "Tlr1"          "Sdc1"          "Ccdc96"       
#> [133] "Ogfod1"        "Pex1"          "Rbm3"          "Dusp15"       
#> [137] "Ahsa2"         "Sgsh"          "Arhgap24"      "Mboat7"       
#> [141] "Ttc26"         "Pacrg"         "AU019823"      "Nip7"         
#> [145] "Lsm11"         "Bmpr2"         "Ctnnal1"       "Zw10"         
#> [149] "Slc31a1"       "St7l"          "Ezh1"          "Zfr2"         
#> [153] "Adam22"        "Mpzl1"         "Etnk1"         "Rpl27a"       
#> [157] "Itsn1"         "Zbtb25"        "Apool"         "Gtf3c4"       
#> [161] "Prss23"        "Elavl1"        "Pop1"          "Sgsm1"        
#> [165] "Ublcp1"        "Usp15"         "Chpt1"         "Fbln1"        
#> [169] "Slit3"         "Kcnip1"        "Hgf"           "Cd84"         
#> [173] "Sft2d2"        "Slc30a10"      "Rasal2"        "Susd4"        
#> [177] "Rsf1"          "Sparcl1"       "Gabrb2"        "Rbck1"        
#> [1] "astrocytes_ependymal"
#> [1] 0.03
#> [1] "Fold enrichment: 1.12063293860355"
#> [1] "Standard deviations from mean: 1.71038221024627"
#> [1] ""
#> [1] "endothelial-mural"
#> [1] 0.03
#> [1] "Fold enrichment: 1.10878298918557"
#> [1] "Standard deviations from mean: 1.619827297495"
#> [1] ""
#> [1] "interneurons"
#> [1] 0.98
#> [1] ""
#> [1] "microglia"
#> [1] 0.06
#> [1] ""
#> [1] "oligodendrocytes"
#> [1] 0.99
#> [1] ""
#> [1] "pyramidal CA1"
#> [1] 0.94
#> [1] ""
#> [1] "pyramidal SS"
#> [1] 0.98
#> [1] ""

# Fill a list with the results
results = add.res.to.merging.list(tt_alzh)
results = add.res.to.merging.list(tt_alzh_BA36,results)
results = add.res.to.merging.list(tt_alzh_BA44,results)