Add to results to merging list

add_res_to_merging_list Adds EWCE results to a list for merging analysis

add_res_to_merging_list(full_res, existing_results = NULL)

Arguments

full_res
results list generated using bootstrap_enrichment_test or ewce_expression_data functions. Multiple results tables can be merged into one results table, as long as the 'list' column is set to distinguish them.

existing_results
Output of previous rounds from adding results to list. Leave empty if this is the first item in the list.

Value

merged results list

Examples

# Load the single cell data
library(ewceData)

#> Loading required package: ExperimentHub

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Loading required package: AnnotationHub

#> Loading required package: BiocFileCache

#> Loading required package: dbplyr

ctd <- ctd()

#> snapshotDate(): 2021-05-18

#> see ?ewceData and browseVignettes('ewceData') for documentation

#> loading from cache

# Load the data
tt_alzh <- tt_alzh()

#> see ?ewceData and browseVignettes('ewceData') for documentation

#> loading from cache

#tt_alzh_BA36 <- tt_alzh_BA36()
# Use 3 bootstrap lists for speed, for publishable analysis use >10000
reps <- 3
# Use 5 up/down regulated genes (thresh) for speed, default is 250
thresh = 5
# Run EWCE analysis
#tt_results <- ewce_expression_data(
#    sct_data = ctd, tt = tt_alzh, annotLevel = 1, thresh = thresh,
#    reps = reps, ttSpecies = "human", sctSpecies = "mouse"
#)
#tt_results_36 <- ewce_expression_data(
#    sct_data = ctd, tt = tt_alzh_BA36, annotLevel = 1, thresh = thresh,
#    reps = reps, ttSpecies = "human", sctSpecies = "mouse"
#)

# Fill a list with the results
results <- add_res_to_merging_list(tt_alzh)
#results <- add_res_to_merging_list(tt_alzh_BA36, results)