bin_columns_into_quantiles

bin_columns_into_quantiles is an internal function used to convert a matrix of specificity (with columns of cell types) intom a matrix of specificity quantiles

bin_columns_into_quantiles(
  matrixIn,
  numberOfBins = 40,
  defaultBin = as.integer(numberOfBins/2)
)

Arguments

matrixIn

The matrix of specificity values

numberOfBins

Number of quantile 'bins' to use (40 is recommended)

defaultBin

Which bin to assign when there's only one non-zero quantile. In situations where there's only one non-zero quantile, cut() throws an error. Avoid these situations by using a default quantile.

Value

A matrix with same shape as matrixIn but with columns storing quantiles instead of specificity

Examples

ctd <- ewceData::ctd()
#> see ?ewceData and browseVignettes('ewceData') for documentation
#> loading from cache
ctd[[1]]$specificity_quantiles <- apply(ctd[[1]]$specificity, 2,
    FUN = bin_columns_into_quantiles,
    numberOfBins = 40
)