Bootstrap celltype enrichment test

bootstrap_enrichment_test takes a genelist and a single cell type transcriptome dataset and determines the probability of enrichment and fold changes for each cell type.

bootstrap_enrichment_test(
  sct_data = NA,
  hits = NA,
  bg = NA,
  genelistSpecies = "mouse",
  sctSpecies = "mouse",
  reps = 100,
  annotLevel = 1,
  geneSizeControl = FALSE,
  controlledCT = NULL
)

Arguments

sct_data
List generated using generate_celltype_data

hits
Array of MGI gene symbols containing the target gene list. Must be HGNC symbols if geneSizeControl=TRUE

bg
Array of MGI gene symbols containing the background gene list. Must be HGNC symbols if geneSizeControl=TRUE

genelistSpecies
Either 'mouse' or 'human' depending on whether MGI or HGNC symbols are used for gene lists

sctSpecies
Either 'mouse' or 'human' depending on whether MGI or HGNC symbols are used for the single cell dataset

reps
Number of random gene lists to generate (default=100 but should be over 10000 for publication quality results)

annotLevel
an integer indicating which level of the annotation to analyse. Default = 1.

geneSizeControl
a logical indicating whether you want to control for GC content and transcript length. Recommended if the gene list originates from genetic studies. Default is FALSE. If set to TRUE then human gene lists should be used rather than mouse.

controlledCT
(optional) If not NULL, and instead is the name of a cell type, then the bootstrapping controls for expression within that cell type

Value

A list containing three data frames:

results: dataframe in which each row gives the statistics (p-value, fold change and number of standard deviations from the mean) associated with the enrichment of the stated cell type in the gene list

hit.cells: vector containing the summed proportion of expression in each cell type for the target list

bootstrap_data: matrix in which each row represents the summed proportion of expression in each cell type for one of the random lists

Examples

library(ewceData)
# Load the single cell data
ctd <- ctd()

#> see ?ewceData and browseVignettes('ewceData') for documentation

#> loading from cache

# Set the parameters for the analysis
# Use 3 bootstrap lists for speed, for publishable analysis use >10000
reps <- 3

# Load the gene list and get human orthologs
example_genelist <- example_genelist()

#> see ?ewceData and browseVignettes('ewceData') for documentation

#> loading from cache

mouse_to_human_homologs <- mouse_to_human_homologs()

#> see ?ewceData and browseVignettes('ewceData') for documentation

#> loading from cache

m2h <- unique(mouse_to_human_homologs[, c("HGNC.symbol", "MGI.symbol")])
mouse.hits <-
    unique(m2h[m2h$HGNC.symbol %in% example_genelist, "MGI.symbol"])
#subset mouse.bg for speed but ensure it still contains the hits
mouse.bg <- unique(c(m2h$MGI.symbol[1:100],mouse.hits))

# Bootstrap significance test, no control for transcript length or GC content
full_results <- bootstrap_enrichment_test(
    sct_data = ctd, hits = mouse.hits,
    bg = mouse.bg, reps = reps, annotLevel = 2, sctSpecies = "mouse",
    genelistSpecies = "mouse"
)

#> see ?ewceData and browseVignettes('ewceData') for documentation

#> loading from cache

#> see ?ewceData and browseVignettes('ewceData') for documentation

#> loading from cache

#> see ?ewceData and browseVignettes('ewceData') for documentation

#> loading from cache

#> Apoe, Inpp5d, Cd2ap, Nme8, Cass4, Mef2c, Zcwpw1, Bin1, Clu, Celf1, Abca7, Slc24a4, Ptk2b, Picalm, Fermt2, Sorl1

#> (none)

#> 0.666666666666667

#> 

#> Astro1

#> 0

#> Fold enrichment: 1.67810589540959

#> Standard deviations from mean: 1.2407003583237

#> 

#> Astro2

#> 0

#> Fold enrichment: 1.79028744888456

#> Standard deviations from mean: 3.75826283824809

#> 

#> CA1Pyr1

#> 0

#> Fold enrichment: 1.34099507539393

#> Standard deviations from mean: 1.7314480601556

#> 

#> CA1Pyr2

#> 0.333333333333333

#> 

#> CA1PyrInt

#> 0

#> Fold enrichment: 1.44133632646825

#> Standard deviations from mean: 1.74816323194772

#> 

#> CA2Pyr2

#> 0.333333333333333

#> 

#> Choroid

#> 0.333333333333333

#> 

#> ClauPyr

#> 1

#> 

#> Epend

#> 0

#> Fold enrichment: 2.28381351191735

#> Standard deviations from mean: 9.63610727998833

#> 

#> Int1

#> 1

#> 

#> Int10

#> 1

#> 

#> Int11

#> 1

#> 

#> Int12

#> 1

#> 

#> Int13

#> 1

#> 

#> Int14

#> 1

#> 

#> Int15

#> 1

#> 

#> Int16

#> 1

#> 

#> Int2

#> 1

#> 

#> Int3

#> 1

#> 

#> Int4

#> 1

#> 

#> Int5

#> 1

#> 

#> Int6

#> 1

#> 

#> Int7

#> 1

#> 

#> Int8

#> 1

#> 

#> Int9

#> 1

#> 

#> Mgl1

#> 0

#> Fold enrichment: 1.9575326363329

#> Standard deviations from mean: 42.5619728625542

#> 

#> Mgl2

#> 0

#> Fold enrichment: 3.73982534374301

#> Standard deviations from mean: 7.08136453023505

#> 

#> Oligo1

#> 0.333333333333333

#> 

#> Oligo2

#> 0.333333333333333

#> 

#> Oligo3

#> 0.333333333333333

#> 

#> Oligo4

#> 0.333333333333333

#> 

#> Oligo5

#> 0.333333333333333

#> 

#> Oligo6

#> 0.333333333333333

#> 

#> Peric

#> 0.333333333333333

#> 

#> Pvm1

#> 0

#> Fold enrichment: 1.94621639203557

#> Standard deviations from mean: 3.65625444006762

#> 

#> Pvm2

#> 0.333333333333333

#> 

#> S1PyrDL

#> 1

#> 

#> S1PyrL23

#> 0.333333333333333

#> 

#> S1PyrL4

#> 0.333333333333333

#> 

#> S1PyrL5

#> 0.333333333333333

#> 

#> S1PyrL5a

#> 0.333333333333333

#> 

#> S1PyrL6

#> 0.333333333333333

#> 

#> S1PyrL6b

#> 0.333333333333333

#> 

#> SubPyr

#> 1

#> 

#> Vend1

#> 1

#> 

#> Vend2

#> 1

#> 

#> Vsmc

#> 0.333333333333333

#>