check.ewce.genelist.inputs

check.ewce.genelist.inputs Is used to check that hits and bg gene lists passed to EWCE are setup correctly. Checks they are the appropriate length. Checks all hits genes are in bg. Checks the species match and if not reduces to 1:1 orthologs.

check.ewce.genelist.inputs(
  sct_data,
  hits,
  bg,
  genelistSpecies,
  sctSpecies,
  geneSizeControl = FALSE
)

Arguments

sct_data
List generated using generate.celltype.data

hits
Array of MGI/HGNC gene symbols containing the target gene list.

bg
Array of MGI/HGNC gene symbols containing the background gene list.

genelistSpecies
Either 'mouse' or 'human' depending on whether MGI or HGNC symbols are used for gene lists

sctSpecies
Either 'mouse' or 'human' depending on whether MGI or HGNC symbols are used for the single cell dataset

geneSizeControl
Will genelists sampled control for GC content and transcript length? Boolean.

Value

A list containing

hits: Array of MGI/HGNC gene symbols containing the target gene list.

bg: Array of MGI/HGNC gene symbols containing the background gene list.

Examples

# Called from "bootstrap.enrichment.test()" and "generate.bootstrap.plots()"
data("ctd")
data("example_genelist")
data("mouse_to_human_homologs")
m2h = unique(mouse_to_human_homologs[,c("HGNC.symbol","MGI.symbol")])
mouse.hits = unique(m2h[m2h$HGNC.symbol %in% example_genelist,"MGI.symbol"])
mouse.bg  = unique(m2h$MGI.symbol)
checkedLists = check.ewce.genelist.inputs(sct_data=ctd,hits=mouse.hits,
    bg=mouse.bg,genelistSpecies="mouse",sctSpecies="mouse")