Create background gene list for multiple species

Create background gene list for the intersection/union between multiple species (gene_list1_species, gene_list2_species, and sctSpecies), and then filter the gene lists to only include genes within the background.

create_background_multilist(
  gene_list1,
  gene_list2,
  gene_list1_species,
  gene_list2_species,
  output_species = "human",
  bg = NULL,
  use_intersect = FALSE,
  method = "homologene",
  verbose = TRUE
)

Arguments

output_species

Species to convert all genes from species1 and species2 to first. Default="human", but can be to either any species supported by orthogene, including species1 or species2.

bg

User supplied background list that will be returned to the user after removing duplicate genes.

use_intersect

When species1 and species2 are both different from output_species, this argument will determine whether to use the intersect (TRUE) or union (FALSE) of all genes from species1 and species2.

method

R package to use for gene mapping:

"gprofiler" : Slower but more species and genes.

"homologene" : Faster but fewer species and genes.

"babelgene" : Faster but fewer species and genes. Also gives consensus scores for each gene mapping based on a several different data sources.

verbose

Print messages.

Value

Background and gene list.