Bootstrap celltype enrichment test for transcriptome data

ewce_expression_data takes a differential expression table and determines the probability of cell-type enrichment in the up & down regulated genes

ewce_expression_data(
  sct_data,
  annotLevel = 1,
  tt,
  sortBy = "t",
  thresh = 250,
  reps = 100,
  ttSpecies = "mouse",
  sctSpecies = "mouse"
)

Arguments

sct_data
List generated using generate.celltype.data

annotLevel
an integer indicating which level of the annotation to analyse. Default = 1.

tt
Differential expression table. Can be output of limma::topTable function. Minimum requirement is that one column stores a metric of increased/decreased expression (i.e. log fold change, t-statistic for differential expression etc) and another contains either HGNC or MGI symbols.

sortBy
Column name of metric in tt which should be used to sort up- from down- regulated genes. Default="t"

thresh
The number of up- and down- regulated genes to be included in each analysis. Dafault=250

reps
Number of random gene lists to generate (default=100 but should be over 10000 for publication quality results)

ttSpecies
Either 'mouse' or 'human' depending on which species the differential expression table was generated from

sctSpecies
Either 'mouse' or 'human' depending on which species the single cell data was generated from

Value

A list containing five data frames:

results: dataframe in which each row gives the statistics (p-value, fold change and number of standard deviations from the mean) associated with the enrichment of the stated cell type in the gene list. An additional column *Direction* stores whether it the result is from the up or downregulated set.

hit.cells.up: vector containing the summed proportion of expression in each cell type for the target list

hit.cells.down: vector containing the summed proportion of expression in each cell type for the target list#'

bootstrap_data.up: matrix in which each row represents the summed proportion of expression in each cell type for one of the random lists

bootstrap_data.down: matrix in which each row represents the summed proportion of expression in each cell type for one of the random lists

Examples

# Load the single cell data
data("ctd")

# Set the parameters for the analysis
reps=100 # <- Use 100 bootstrap lists so it runs quickly, for publishable analysis use >10000
annotLevel=1 # <- Use cell level annotations (i.e. Interneurons)

# Load the top table
data(tt_alzh)

tt_results = ewce_expression_data(sct_data=ctd,tt=tt_alzh,annotLevel=1,
    ttSpecies="human",sctSpecies="mouse")

#>   [1] "Myom1"         "Rbm17"         "Max"           "Mbd3"         
#>   [5] "Wdr33"         "Acss1"         "Pcyt1b"        "Bcam"         
#>   [9] "Prpf38b"       "Ppp2r5c"       "Ica1"          "Cobl"         
#>  [13] "Marveld1"      "Gemin4"        "Zfp668"        "Casc3"        
#>  [17] "Aff3"          "Tcf3"          "Taf1c"         "Gna12"        
#>  [21] "Lamp2"         "Inpp5d"        "Phf14"         "Stk11"        
#>  [25] "Plxnd1"        "Aebp1"         "Pgf"           "5430427O19Rik"
#>  [29] "Gad1"          "App"           "Casp9"         "Mid1ip1"      
#>  [33] "Cep70"         "Dnajc1"        "Cbx5"          "Llph"         
#>  [37] "Rbm6"          "Tbc1d2b"       "Csk"           "Col6a1"       
#>  [41] "1110034G24Rik" "Vars2"         "Lama2"         "Aup1"         
#>  [45] "Ctbp2"         "Atg3"          "Aifm3"         "Gga1"         
#>  [49] "Zcchc24"       "Prtg"          "Ralgps1"       "Ncan"         
#>  [53] "Otud4"         "Phf21a"        "Adra2c"        "Iqcc"         
#>  [57] "Ptbp1"         "Col27a1"       "Zfp526"        "Nrp2"         
#>  [61] "Flnb"          "Dcn"           "Zfp36l1"       "Antxr1"       
#>  [65] "Plod3"         "Ddit4l"        "Hnrnph3"       "Apba3"        
#>  [69] "Taf5"          "Mmd2"          "Kif1c"         "Zswim1"       
#>  [73] "Myh14"         "Wwox"          "Tmem47"        "Zfp36l2"      
#>  [77] "Palld"         "Mfge8"         "Nfic"          "Ilf3"         
#>  [81] "Armc3"         "Myh11"         "Ppp4r1"        "Plekha8"      
#>  [85] "Gas2l1"        "Tal1"          "Atp8b2"        "Slc40a1"      
#>  [89] "Enah"          "1110059E24Rik" "Nsd1"          "Rprm"         
#>  [93] "Golim4"        "Paox"          "Numa1"         "Rabep1"       
#>  [97] "Tcf12"         "Fastk"         "Cnot4"         "Cxcr4"        
#> [101] "Rrad"          "Dmc1"          "Rassf9"        "Abca9"        
#> [105] "Hbp1"          "Klk7"          "Nedd4"         "Adck1"        
#> [109] "Smarcc1"       "Ezr"           "Gpr161"        "Rapgef3"      
#> [113] "Snrnp48"       "Abi3bp"        "Nav1"          "Ssr1"         
#> [117] "Pknox1"        "Qk"            "Osbp2"         "Col5a3"       
#> [121] "Fgfr1"         "Repin1"        "Dlgap4"        "Atp5sl"       
#> [125] "Rimkla"        "Pdcd2"         "Alcam"         "P2rx7"        
#> [129] "Smyd4"         "Mtss1l"        "Sfxn5"         "Unc5b"        
#> [133] "Ddc"           "Pcdh1"         "Igf2"          "Sec11a"       
#> [137] "Osbpl2"        "3110079O15Rik" "Sall3"         "Git1"         
#> [141] "Llgl2"         "Zfp385a"       "Kank2"         "Rps21"        
#> [145] "Smg5"          "Gnao1"         "Csnk1d"        "Ubxn1"        
#> [149] "Sos1"          "Ascl1"         "Map2k3"        "Nf2"          
#> [153] "Fbxw4"         "Ocln"          "Zc3h14"        "S100a4"       
#> [157] "Rftn2"         "Zfp444"        "Rab28"         "Notch2"       
#> [161] "Efcab2"        "Maf"           "Hes5"          "Chdh"         
#> [165] "Mdm2"          "Arhgap22"      "Rnf11"         "Crtap"        
#> [169] "Frzb"          "Gja1"          "Lmna"          "Rheb"         
#> [173] "Tmem184b"      "Shroom1"       "Ahnak"         "Itsn1"        
#> [177] "Jmjd6"         "Kcnj10"        "Bgn"           "Atg12"        
#> [181] "Ngfr"          "Slc25a29"      "Pcmtd1"        "Adam33"       
#> [185] "Rnd2"          "5031439G07Rik" "Mif4gd"        "Atp6v0e"      
#> [189] "Hist3h2a"      "Tmcc2"         "Mrgprf"        "Nrip2"        
#> [193] "Lef1"          "Sox12"         "Med25"         "Jund"         
#> [197] "Acp6"          "Ak1"           "Cd79b"         "Wwc1"         
#> [201] "Itga7"         "Map3k11"       "Cspg4"         "Vat1"         
#> [205] "Zfp692"        "Selplg"        "Dbi"           "Pou3f2"       
#> [209] "Doc2b"         "Itgb5"         "Polr3g"        "Pacs1"        
#> [1] "astrocytes_ependymal"
#> [1] 0
#> [1] "Fold enrichment: 1.23472112735846"
#> [1] "Standard deviations from mean: 3.70467623557712"
#> [1] ""
#> [1] "endothelial-mural"
#> [1] 0
#> [1] "Fold enrichment: 1.25586447507816"
#> [1] "Standard deviations from mean: 4.00410843641787"
#> [1] ""
#> [1] "interneurons"
#> [1] 1
#> [1] ""
#> [1] "microglia"
#> [1] 0.02
#> [1] "Fold enrichment: 1.19795554625304"
#> [1] "Standard deviations from mean: 2.49320466000033"
#> [1] ""
#> [1] "oligodendrocytes"
#> [1] 0.78
#> [1] ""
#> [1] "pyramidal CA1"
#> [1] 1
#> [1] ""
#> [1] "pyramidal SS"
#> [1] 1
#> [1] ""
#>   [1] "Prmt5"         "Atg7"          "Ivd"           "Ica1"         
#>   [5] "Farsb"         "Aifm1"         "Pole3"         "Psmd11"       
#>   [9] "Dpm2"          "Vrk3"          "Dnajc5"        "Gphn"         
#>  [13] "Tubb5"         "Ets2"          "Pygb"          "Fgf12"        
#>  [17] "Me2"           "Tfrc"          "Adcy3"         "Ramp3"        
#>  [21] "Cops7a"        "Snrpb"         "Retsat"        "Reep1"        
#>  [25] "Nap1l5"        "Rufy3"         "Psmg1"         "Dclk1"        
#>  [29] "Carf"          "Vsnl1"         "Sec23a"        "Liph"         
#>  [33] "Pias2"         "Kif9"          "Hpcal1"        "Med14"        
#>  [37] "Lmo4"          "Ank"           "Col12a1"       "Lrrc7"        
#>  [41] "Adam23"        "Lzic"          "Fgf14"         "Tollip"       
#>  [45] "Gnal"          "Lpin2"         "Rsrc1"         "Necap1"       
#>  [49] "Atp6v1b2"      "Kdm5c"         "Coq6"          "Rhbdd2"       
#>  [53] "Mef2c"         "Gabrd"         "Ajap1"         "Pcbp2"        
#>  [57] "Brwd1"         "Atp6v1g1"      "Mysm1"         "H2afy"        
#>  [61] "Parp2"         "Fam175b"       "Mad2l1bp"      "Aak1"         
#>  [65] "Uqcrc2"        "Rragb"         "Cd200"         "Phtf1"        
#>  [69] "Psd3"          "Rer1"          "Grin2a"        "Atp6ap2"      
#>  [73] "Slc27a2"       "Ddx19a"        "Cirbp"         "Camk1g"       
#>  [77] "Srp72"         "Zfp697"        "Cdc5l"         "Prkag2"       
#>  [81] "Pdk2"          "Gng3"          "Ddx10"         "Farsa"        
#>  [85] "Nrxn3"         "Hnrnpd"        "Pgs1"          "Fgfr1op"      
#>  [89] "Nrn1"          "Commd4"        "Msi2"          "Mapk10"       
#>  [93] "Npc1"          "Cap1"          "Pdss1"         "Eif4b"        
#>  [97] "Pgk1"          "Stx18"         "Slc20a1"       "Gprasp1"      
#> [101] "Atp5b"         "Ppp2ca"        "Zcchc17"       "Tubgcp4"      
#> [105] "Dlat"          "Ypel1"         "Rab2a"         "Gns"          
#> [109] "Snip1"         "Thumpd3"       "Pcyox1l"       "Pgm2l1"       
#> [113] "Gdap1"         "Nat10"         "Arih2"         "Prkab1"       
#> [117] "Rgs4"          "Asap1"         "Amph"          "Prep"         
#> [121] "Kcnq3"         "Sema3a"        "G6pc3"         "Wdr1"         
#> [125] "Nek3"          "Ndufa10"       "Vamp1"         "Prkcd"        
#> [129] "Fosb"          "Mtpap"         "Lgals8"        "Rnf41"        
#> [133] "Igf1"          "Krba1"         "Pfkm"          "Slc35a3"      
#> [137] "Stmn2"         "Scg3"          "Psma3"         "Snx17"        
#> [141] "Ephx2"         "Atp8a2"        "Rfc2"          "Crls1"        
#> [145] "Daglb"         "Gabra4"        "Dennd3"        "Jak3"         
#> [149] "Ogfod1"        "Pex1"          "Tmem59"        "Rbm3"         
#> [153] "Sgsh"          "Bcl2l13"       "Sdc3"          "Stat4"        
#> [157] "Mlx"           "Kctd5"         "Fads3"         "Trappc2l"     
#> [161] "Tbc1d9"        "Ptk2b"         "Thnsl2"        "Txn2"         
#> [165] "Mcph1"         "Snx10"         "Hs3st2"        "Camk2g"       
#> [169] "Npat"          "Kdelr2"        "Xpnpep1"       "Axin2"        
#> [173] "Mak16"         "Cacnb2"        "Prpf4"         "Zfp398"       
#> [177] "Wars"          "Cep57"         "Eif5"          "Ndfip1"       
#> [181] "Ndrg4"         "Nol6"          "Letmd1"        "Prss23"       
#> [185] "Gaa"           "Bicd2"         "Dhrs7b"        "M6pr"         
#> [189] "Usp11"         "Cdc37l1"       "Map2k5"        "Ywhaz"        
#> [193] "Fen1"          "Atp6ap1"       "Nedd4l"        "Mcam"         
#> [197] "Vps53"         "Slc17a7"       "Cobll1"        "Actr1a"       
#> [201] "Tiam2"         "Mrps18b"       "Slc38a7"       "Gabrg2"       
#> [205] "Pex3"          "Pdcd4"         "4930430F08Rik" "Lrrtm2"       
#> [209] "Cadps2"        "Mybbp1a"       "Ttc19"         "Lrrfip1"      
#> [213] "Susd4"         "Mat2b"         "Cyb561"        "Snap47"       
#> [217] "Sssca1"        "Tomm20"       
#> [1] "astrocytes_ependymal"
#> [1] 1
#> [1] ""
#> [1] "endothelial-mural"
#> [1] 1
#> [1] ""
#> [1] "interneurons"
#> [1] 0
#> [1] "Fold enrichment: 1.23048391288061"
#> [1] "Standard deviations from mean: 4.0669438085213"
#> [1] ""
#> [1] "microglia"
#> [1] 1
#> [1] ""
#> [1] "oligodendrocytes"
#> [1] 0.99
#> [1] ""
#> [1] "pyramidal CA1"
#> [1] 0
#> [1] "Fold enrichment: 1.21581196595115"
#> [1] "Standard deviations from mean: 4.82508404210989"
#> [1] ""
#> [1] "pyramidal SS"
#> [1] 0
#> [1] "Fold enrichment: 1.29366199557388"
#> [1] "Standard deviations from mean: 5.99250392952273"
#> [1] ""