Example bootstrap celltype enrichment test for transcriptome data

Example celltype enrichment results produced by ewce_expression_data.

example_transcriptome_results(verbose = TRUE, localHub = FALSE)

Source

## Load the single cell data

ctd <- ewceData::ctd()

## Set the parameters for the analysis

## Use 3 bootstrap lists for speed, for publishable analysis use >10,000

reps <- 3

annotLevel <- 1 # <- Use cell level annotations (i.e. Interneurons)

## Use 5 up/down regulated genes (thresh) for speed, default is 250

thresh <- 5

## Load the top table

tt_alzh <- ewceData::tt_alzh()

tt_results <- EWCE::ewce_expression_data( sct_data = ctd, tt = tt_alzh, annotLevel = 1, thresh = thresh, reps = reps, ttSpecies = "human", sctSpecies = "mouse" )

save(tt_results, file = "inst/extdata/tt_results.rda")

Arguments

verbose

Print messages.

localHub

If working offline, add argument localHub=TRUE to work with a local, non-updated hub; It will only have resources available that have previously been downloaded. If offline, Please also see BiocManager vignette section on offline use to ensure proper functionality.

Value

List with 5 items.

Examples

tt_results <- EWCE::example_transcriptome_results()
#> Loading precomputed example transcriptome results.