Extract a matrix from a CellTypeDataset

Extracts a particular matrix (e.g., mean_exp, specificity) from a CellTypeDataset object.

extract_matrix(
  ctd,
  dataset,
  level = 1,
  input_species = NULL,
  output_species = "human",
  metric = "specificity",
  non121_strategy = "drop_both_species",
  numberOfBins = 40,
  remove_unlabeled_clusters = FALSE,
  force_new_quantiles = FALSE,
  as_sparse = TRUE,
  as_DelayedArray = FALSE,
  rename_columns = FALSE,
  verbose = TRUE
)

Arguments

ctd
Input CellTypeData.

dataset
CellTypeData. name.

level
CTD level to extract from.

input_species
Which species the gene names in exp come from.

output_species
Which species' genes names to convert exp to.

metric
Name of the matrix to extract.

non121_strategy
How to handle genes that don't have 1:1 mappings between input_species:output_species. Options include:

"drop_both_species" or "dbs" or 1 :
Drop genes that have duplicate mappings in either the input_species or output_species
(DEFAULT).

"drop_input_species" or "dis" or 2 :
Only drop genes that have duplicate mappings in the input_species.

"drop_output_species" or "dos" or 3 :
Only drop genes that have duplicate mappings in the output_species.

"keep_both_species" or "kbs" or 4 :
Keep all genes regardless of whether they have duplicate mappings in either species.

"keep_popular" or "kp" or 5 :
Return only the most "popular" interspecies ortholog mappings. This procedure tends to yield a greater number of returned genes but at the cost of many of them not being true biological 1:1 orthologs.

"sum","mean","median","min" or "max" :
When gene_df is a matrix and gene_output="rownames", these options will aggregate many-to-one gene mappings (input_species-to-output_species) after dropping any duplicate genes in the output_species.

numberOfBins
Number of non-zero quantile bins.

remove_unlabeled_clusters
Remove any samples that have numeric column names.

force_new_quantiles
By default, quantile computation is skipped if they have already been computed. Set =TRUE to override this and generate new quantiles.

as_sparse
Convert to sparse matrix.

as_DelayedArray
Convert to DelayedArray.

rename_columns
Add species and dataset to column names.

verbose
Print messages.

Value

(specificity) matrix.