Generate bootstrap plots

generate.bootstrap.plots takes a genelist and a single cell type transcriptome dataset and generates plots which show how the expression of the genes in the list compares to those in randomly generated gene lists

generate.bootstrap.plots(
  sct_data,
  hits,
  bg,
  genelistSpecies = "mouse",
  sctSpecies = "mouse",
  reps,
  annotLevel = 1,
  full_results = NA,
  listFileName = "",
  savePath = "~/"
)

Arguments

sct_data
List generated using generate.celltype.data

hits
Array of MGI/HGNC gene symbols containing the target gene list.

bg
Array of MGI/HGNC gene symbols containing the background gene list.

genelistSpecies
Either 'mouse' or 'human' depending on whether MGI or HGNC symbols are used for gene lists

sctSpecies
Either 'mouse' or 'human' depending on whether MGI or HGNC symbols are used for the single cell dataset

reps
Number of random gene lists to generate (default=100 but should be over 10000 for publication quality results)

annotLevel
an integer indicating which level of the annotation to analyse. Default = 1.

full_results
The full output of bootstrap.enrichment.test for the same genelist

listFileName
String used as the root for files saved using this function

savePath
Directory where the BootstrapPlots folder should be saved

Value

Saves a set of pdf files containing graphs. These will be saved with the filename adjusted using the value of listFileName. The files are saved into the 'BootstrapPlot' folder. The files start with one of the following:

qqplot_noText: sorts the gene list according to how enriched it is in the relevant celltype. Plots the value in the target list against the mean value in the bootstrapped lists.

qqplot_wtGSym: as above but labels the gene symbols for the highest expressed genes.

bootDists: rather than just showing the mean of the bootstrapped lists, a boxplot shows the distribution of values

bootDists_LOG: shows the bootstrapped distributions with the y-axis shown on a log scale

Examples

# Load the single cell data
data("ctd")

# Set the parameters for the analysis
reps=100 # <- Use 100 bootstrap lists so it runs quickly, for publishable analysis use >10000

# Load the gene list and get human orthologs
data("example_genelist")
data("mouse_to_human_homologs")
m2h = unique(mouse_to_human_homologs[,c("HGNC.symbol","MGI.symbol")])
mouse.hits = unique(m2h[m2h$HGNC.symbol %in% example_genelist,"MGI.symbol"])
human.hits = unique(m2h[m2h$HGNC.symbol %in% example_genelist,"HGNC.symbol"])
human.bg = unique(m2h$HGNC.symbol)
mouse.bg  = unique(m2h$MGI.symbol)

# Bootstrap significance testing, without controlling for transcript length and GC content
full_results = bootstrap.enrichment.test(sct_data=ctd,hits=mouse.hits,
  bg=mouse.bg,reps=reps,annotLevel=1,sctSpecies="mouse",genelistSpecies="mouse")

#>  [1] "Apoe"    "Inpp5d"  "Cd2ap"   "Nme8"    "Cass4"   "Mef2c"   "Zcwpw1" 
#>  [8] "Bin1"    "Clu"     "Celf1"   "Abca7"   "Slc24a4" "Ptk2b"   "Picalm" 
#> [15] "Fermt2"  "Sorl1"  
#> [1] "astrocytes_ependymal"
#> [1] 0.03
#> [1] "Fold enrichment: 1.40493524893616"
#> [1] "Standard deviations from mean: 1.93113566371208"
#> [1] ""
#> [1] "endothelial-mural"
#> [1] 0.75
#> [1] ""
#> [1] "interneurons"
#> [1] 1
#> [1] ""
#> [1] "microglia"
#> [1] 0.02
#> [1] "Fold enrichment: 1.56701608190953"
#> [1] "Standard deviations from mean: 2.29588307920279"
#> [1] ""
#> [1] "oligodendrocytes"
#> [1] 0.81
#> [1] ""
#> [1] "pyramidal CA1"
#> [1] 0.57
#> [1] ""
#> [1] "pyramidal SS"
#> [1] 0.75
#> [1] ""

generate.bootstrap.plots(sct_data=ctd,hits=mouse.hits, bg=mouse.bg,
     reps=reps,full_results=full_results,listFileName="Example",
     genelistSpecies="mouse",sctSpecies="mouse", annotLevel=1)

#> Warning: Transformation introduced infinite values in continuous y-axis

#> Warning: Transformation introduced infinite values in continuous y-axis

#> Warning: Transformation introduced infinite values in continuous y-axis

#> Warning: Transformation introduced infinite values in continuous y-axis