Run DGE: MAST

Run Differential Gene Expression with MAST.

run_mast(exp, level2annot, test = "LRT", mtc_method = "BH", no_cores = 1, ...)

Source

MAST tutorial

Arguments

exp

Expression matrix with gene names as rownames.

level2annot

Array of cell types, with each sequentially corresponding a column in the expression matrix.

mtc_method

Multiple-testing correction method used by DGE step. See p.adjust for more details.

no_cores

Number of cores to parallelise DGE across.

...

Additional arguments to be passed to gorth or homologene.

NOTE: To return only the most "popular" interspecies ortholog mappings, supply mthreshold=1 here AND set method="gprofiler" above. This procedure tends to yield a greater number of returned genes but at the cost of many of them not being true biological 1:1 orthologs.

For more details, please see here.

Value

MAST results