merge_two_expfiles

merge_two_expfiles Used to combine two single cell type datasets

merge_two_expfiles(exp1, exp2, annot1, annot2, name1 = "", name2 = "")

Arguments

exp1
Numerical expression matrix for dataset1 with row for each gene and column for each cell. Row names are MGI/HGNC gene symbols. Column names are cell IDs which can be cross referenced against the annot data frame.

exp2
Numerical expression matrix for dataset2 with row for each gene and column for each cell. Row names are MGI/HGNC gene symbols. Column names are cell IDs which can be cross referenced against the annot data frame.

annot1
Annotation data frame for dataset1 which contains three columns at least: cell_id, level1class and level2class

annot2
Annotation data frame for dataset2 which contains three columns at least: cell_id, level1class and level2class

name1
Name used to refer to dataset 1. Leave blank if it's already a merged dataset.

name2
Name used to refer to dataset 2. Leave blank if it's already a merged dataset.

Value

List containing merged exp and annot

Examples

# See the EWCE vignette